feat: added unlike character feature for all views
